src/test/java/apoc/agg/CollTest.java
package apoc.agg;

import apoc.util.TestUtil;
import org.junit.AfterClass;
import org.junit.BeforeClass;
import org.junit.Test;
import org.neo4j.graphdb.GraphDatabaseService;
import org.neo4j.test.TestGraphDatabaseFactory;

import java.util.*;

import static apoc.util.TestUtil.testCall;
import static apoc.util.TestUtil.testResult;
import static java.util.Arrays.asList;
import static org.junit.Assert.assertEquals;

public class CollTest {

    private static GraphDatabaseService db;

    @BeforeClass public static void setUp() throws Exception {
        db = new TestGraphDatabaseFactory().newImpermanentDatabase();
        TestUtil.registerProcedure(db, Coll.class);
    }

    @AfterClass public static void tearDown() {
        db.shutdown();
    }

    @Test
    public void testIntersection() throws Exception {
        testCall(db, "UNWIND range(1,2) AS i RETURN apoc.agg.intersection(range(0,i)) AS value",
                (row) -> {
                    Object value = row.get("value");
                    List<Long> expected = asList(0L, 1L);
                    assertEquals(expected, value);
                });
    }

}
